django/hbec-fof-algorithm-service/requirement.txt
scipy==1.2.1
matplotlib==3.0.3
# pymongo==3.7.2
numpy==1.16.2
pandas==0.24.2
statsmodels==0.9.0
djangorestframework==3.9.2
Python==3.7.3
django==2.0.5
simplejson==3.16.0
mysqlclient==1.3.14
django-rest-swagger==2.2.0
bunch==1.0.1
# mock==2.0.0
# redis-py
django-redis
# pgdb==0.0.8
DBUtils
scikit-learn
